fastapi-app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
azure-storage-blob
aiohttp
aiofiles
python-multipart
pydantic>=2.0.0,<3.0.0
psycopg2-binary
python-dotenv
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt] 